requirements.txt
flask
werkzeug
gunicorn
flask_sqlalchemy
pymysql
cryptography
dotenv